+++ sickrage/core/version_updater.py
@@ -447,7 +447,7 @@ class GitUpdateManager(UpdateManager):
                                                                                    self.current_branch))
         if exit_status == 0:
             sickrage.app.log.info("Updating SiCKRAGE from GIT servers")
-            Notifiers.notify_version_update(self.get_newest_version)
+            Notifiers.mass_notify_version_update(self.get_newest_version)
             self.install_requirements()
             return True
 
@@ -646,7 +646,7 @@ class SourceUpdateManager(UpdateManager):
             return False
 
         # Notify update successful
-        Notifiers.notify_version_update(self.get_newest_version)
+        Notifiers.mass_notify_version_update(self.get_newest_version)
 
         # install requirements
         self.install_requirements()
@@ -721,7 +721,7 @@ class PipUpdateManager(UpdateManager):
         __, __, exit_status = self._pip_cmd(self._pip_path, 'install -U --no-cache-dir sickrage')
         if exit_status == 0:
             sickrage.app.log.info("Updating SiCKRAGE from PyPi servers")
-            Notifiers.notify_version_update(self.get_newest_version)
+            Notifiers.mass_notify_version_update(self.get_newest_version)
             return True
 
         return False