+++ sickrage/notifiers/boxcar2.py
@@ -85,28 +85,28 @@ class Boxcar2Notifier(Notifiers):
         sickrage.app.log.debug("Boxcar2 notification successful.")
         return True
 
-    def _notify_snatch(self, ep_name, title=None):
+    def notify_snatch(self, ep_name, title=None):
         if not title:
             title = self.notifyStrings[self.NOTIFY_SNATCH]
 
         if sickrage.app.config.boxcar2_notify_onsnatch:
             self._notifyBoxcar2(title, ep_name)
 
-    def _notify_download(self, ep_name, title=None):
+    def notify_download(self, ep_name, title=None):
         if not title:
             title = self.notifyStrings[self.NOTIFY_DOWNLOAD]
 
         if sickrage.app.config.boxcar2_notify_ondownload:
             self._notifyBoxcar2(title, ep_name)
 
-    def _notify_subtitle_download(self, ep_name, lang, title=None):
+    def notify_subtitle_download(self, ep_name, lang, title=None):
         if not title:
             title = self.notifyStrings[self.NOTIFY_SUBTITLE_DOWNLOAD]
 
         if sickrage.app.config.boxcar2_notify_onsubtitledownload:
             self._notifyBoxcar2(title, ep_name + ": " + lang)
 
-    def _notify_version_update(self, new_version="??"):
+    def notify_version_update(self, new_version="??"):
         if sickrage.app.config.use_boxcar2:
             update_text = self.notifyStrings[self.NOTIFY_GIT_UPDATE_TEXT]
             title = self.notifyStrings[self.NOTIFY_GIT_UPDATE]