+++ sickrage/notifiers/emby.py
@@ -31,19 +31,19 @@ class EMBYNotifier(Notifiers):
         super(EMBYNotifier, self).__init__()
         self.name = 'emby'
 
-    def _notify_snatch(self, ep_name):
+    def notify_snatch(self, ep_name):
         if sickrage.app.config.emby_notify_onsnatch:
             self._notify_emby(self.notifyStrings[self.NOTIFY_SNATCH] + ': ' + ep_name)
 
-    def _notify_download(self, ep_name):
+    def notify_download(self, ep_name):
         if sickrage.app.config.emby_notify_ondownload:
             self._notify_emby(self.notifyStrings[self.NOTIFY_DOWNLOAD] + ': ' + ep_name)
 
-    def _notify_subtitle_download(self, ep_name, lang):
+    def notify_subtitle_download(self, ep_name, lang):
         if sickrage.app.config.emby_notify_onsubtitledownload:
             self._notify_emby(self.notifyStrings[self.NOTIFY_SUBTITLE_DOWNLOAD] + ' ' + ep_name + ": " + lang)
 
-    def _notify_version_update(self, new_version="??"):
+    def notify_version_update(self, new_version="??"):
         if sickrage.app.config.use_emby:
             update_text = self.notifyStrings[self.NOTIFY_GIT_UPDATE_TEXT]
             title = self.notifyStrings[self.NOTIFY_GIT_UPDATE]
@@ -86,7 +86,7 @@ class EMBYNotifier(Notifiers):
     def test_notify(self, host, emby_apikey):
         return self._notify_emby('This is a test notification from SiCKRAGE', host, emby_apikey)
 
-    def notify_login(self, ipaddress=""):
+    def mass_notify_login(self, ipaddress=""):
         if sickrage.app.config.use_emby:
             update_text = self.notifyStrings[self.NOTIFY_LOGIN_TEXT]
             title = self.notifyStrings[self.NOTIFY_LOGIN]