+++ sickrage/notifiers/twilio_notifer.py
@@ -46,24 +46,24 @@ class TwilioNotifier(Notifiers):
         return TwilioRestClient(sickrage.app.config.twilio_account_sid,
                                 sickrage.app.config.twilio_auth_token)
 
-    def _notify_snatch(self, ep_name):
+    def notify_snatch(self, ep_name):
         if sickrage.app.config.twilio_notify_onsnatch:
             self._notifyTwilio(self.notifyStrings[self.NOTIFY_SNATCH] + ': ' + ep_name)
 
-    def _notify_download(self, ep_name):
+    def notify_download(self, ep_name):
         if sickrage.app.config.twilio_notify_ondownload:
             self._notifyTwilio(self.notifyStrings[self.NOTIFY_DOWNLOAD] + ': ' + ep_name)
 
-    def _notify_subtitle_download(self, ep_name, lang):
+    def notify_subtitle_download(self, ep_name, lang):
         if sickrage.app.config.twilio_notify_onsubtitledownload:
             self._notifyTwilio(self.notifyStrings[self.NOTIFY_SUBTITLE_DOWNLOAD] + ' ' + ep_name + ': ' + lang)
 
-    def _notify_version_update(self, new_version):
+    def notify_version_update(self, new_version):
         if sickrage.app.config.use_twilio:
             update_text = self.notifyStrings[self.NOTIFY_GIT_UPDATE_TEXT]
             self._notifyTwilio(update_text + new_version)
 
-    def _notify_login(self, ipaddress=""):
+    def notify_login(self, ipaddress=""):
         if sickrage.app.config.use_twilio:
             update_text = self.notifyStrings[self.NOTIFY_LOGIN_TEXT]
             title = self.notifyStrings[self.NOTIFY_LOGIN]